pebble-challtestsrv: deleted
I spent a lot of time with this, but it turned out that pebble itself is
limited to making A/AAAA lookups with specified DNS servers, and
pebble-challtestsrv DNS functionality is only able to provide fixed
responses and not able to resolve an IP all the way using a CNAME for
example.

So, instead of keeping it around, I'm deleting it to reduce complexity
that I don't know if it make sense to have in this Helm chart meant not
to test acme-clients themselves, but the integration of acme-clients in
applications against an ACME server.