Add basic values.schema.json, a helm3 feature
I've read about schema files for values.yaml, which are applied as part
of helm install/upgrade/template/lint which is quite nice. This schema
does not cover all the configuration, and I don't think it is worth to
do that, but I figure it is a good trial of the feature.

ref: https://helm.sh/docs/topics/charts/#schema-files